requirements.txt
discord.py[voice]
PyNaCl
youtube-dl
soundcloud
pygame
ics